requirements.txt
Flask
pandas
Pillow
PyPDF2
pdf2docx
pdfplumber
